# openapi-spec.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/v1/visa/{id}": {
            "get": {
                "tags": [
                    "visa-controller"
                ],
                "operationId": "getVisaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visa"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "visa-controller"
                ],
                "operationId": "updateVisa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Visa"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visa"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "visa-controller"
                ],
                "operationId": "deleteVisa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/visa/savePotentialEmployees": {
            "post": {
                "tags": [
                    "visa-controller"
                ],
                "operationId": "savePotentialEmployees",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/PotentialEmployee"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PotentialEmployee"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/visa/potentialEmployees": {
            "post": {
                "tags": [
                    "visa-controller"
                ],
                "operationId": "createPotentialEmployee",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PotentialEmployee"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PotentialEmployee"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/visa/createVisa": {
            "post": {
                "tags": [
                    "visa-controller"
                ],
                "operationId": "createVisa",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Visa"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visa"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/visa/employee/{employeeId}": {
            "get": {
                "tags": [
                    "visa-controller"
                ],
                "operationId": "getVisasByEmployee",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Visa"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/visa/employee/{employeeId}/latest": {
            "get": {
                "tags": [
                    "visa-controller"
                ],
                "operationId": "getLatestVisaByEmployee",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visa"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Address": {
                "type": "object",
                "properties": {
                    "addressId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "address": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "zipCode": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "employee": {
                        "$ref": "#/components/schemas/Employee"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Assignment": {
                "type": "object",
                "properties": {
                    "assignmentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "projectId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "assignmentType": {
                        "type": "string"
                    },
                    "assignmentTaxType": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "wage": {
                        "type": "number",
                        "format": "float"
                    },
                    "status": {
                        "type": "string"
                    },
                    "employeeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "description": {
                        "type": "string"
                    },
                    "assignmentNotes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Notes"
                        }
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Customer": {
                "type": "object",
                "properties": {
                    "customerId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "customerCompanyName": {
                        "type": "string"
                    },
                    "customerName": {
                        "type": "string"
                    },
                    "customerPhone": {
                        "type": "string"
                    },
                    "customerEmail": {
                        "type": "string"
                    },
                    "ein": {
                        "type": "string"
                    },
                    "website": {
                        "type": "string"
                    },
                    "customerStatus": {
                        "type": "string"
                    },
                    "customerStartDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "customerEndDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "customerContactEmail": {
                        "type": "string"
                    },
                    "customerType": {
                        "type": "string"
                    },
                    "customerAddress": {
                        "type": "string"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Employee": {
                "type": "object",
                "properties": {
                    "employeeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "emailId": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "dob": {
                        "type": "string"
                    },
                    "ssn": {
                        "type": "string"
                    },
                    "taxTerm": {
                        "type": "string"
                    },
                    "referredBy": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "designation": {
                        "type": "string"
                    },
                    "employmentType": {
                        "type": "string"
                    },
                    "employeeType": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "primarySkills": {
                        "type": "string"
                    },
                    "secondarySkills": {
                        "type": "string"
                    },
                    "workCity": {
                        "type": "string"
                    },
                    "workCountry": {
                        "type": "string"
                    },
                    "resourceType": {
                        "type": "string"
                    },
                    "employeeDept": {
                        "type": "string"
                    },
                    "annualPay": {
                        "type": "number",
                        "format": "float"
                    },
                    "payrollStart": {
                        "type": "string",
                        "format": "date"
                    },
                    "everifyStatus": {
                        "type": "string"
                    },
                    "insurance": {
                        "type": "string"
                    },
                    "companyName": {
                        "type": "string"
                    },
                    "address": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Address"
                        }
                    },
                    "project": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Project"
                        }
                    },
                    "visas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Visa"
                        }
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    },
                    "i9": {
                        "type": "string"
                    },
                    "paf": {
                        "type": "string"
                    }
                }
            },
            "LCA": {
                "type": "object",
                "properties": {
                    "lcaId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "jobTitle": {
                        "type": "string"
                    },
                    "lcaCaseNumber": {
                        "type": "string"
                    },
                    "employmentStartDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "employmentEndDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "lcaPostedFromDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "lcaPostedToDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "socCode": {
                        "type": "string"
                    },
                    "lcaWage": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "mangerId": {
                        "type": "string"
                    },
                    "jobLocation": {
                        "type": "string"
                    },
                    "jobLocation2": {
                        "type": "string"
                    },
                    "noticePostedLocation": {
                        "type": "string"
                    },
                    "noticePostedLocation2": {
                        "type": "string"
                    },
                    "lcaNumber": {
                        "type": "string"
                    },
                    "client": {
                        "type": "string"
                    },
                    "vendor": {
                        "type": "string"
                    },
                    "certifiedDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "type": "string"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Notes": {
                "type": "object",
                "properties": {
                    "noteId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "notesType": {
                        "type": "string"
                    },
                    "details": {
                        "type": "string"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Project": {
                "type": "object",
                "properties": {
                    "projectId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "projectName": {
                        "type": "string"
                    },
                    "customer": {
                        "$ref": "#/components/schemas/Customer"
                    },
                    "employee": {
                        "$ref": "#/components/schemas/Employee"
                    },
                    "client": {
                        "type": "string"
                    },
                    "billRates": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Wage"
                        }
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "invoiceTerm": {
                        "type": "string"
                    },
                    "paymentTerm": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "assignments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Assignment"
                        }
                    },
                    "employeeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "vendorId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Visa": {
                "type": "object",
                "properties": {
                    "visaId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "visaCategory": {
                        "type": "string"
                    },
                    "receiptNumber": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "jobTitle": {
                        "type": "string"
                    },
                    "lcaNumber": {
                        "type": "string"
                    },
                    "socCode": {
                        "type": "string"
                    },
                    "client": {
                        "type": "string"
                    },
                    "vendor": {
                        "type": "string"
                    },
                    "jobLocation": {
                        "type": "string"
                    },
                    "jobLocation2": {
                        "type": "string"
                    },
                    "lcaWage": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "status": {
                        "type": "string"
                    },
                    "filingType": {
                        "type": "string",
                        "enum": [
                            "TRANSFER",
                            "CONSULAR",
                            "COS"
                        ]
                    },
                    "visaSubCategory": {
                        "type": "string",
                        "enum": [
                            "TRANSFER",
                            "CAP",
                            "TRANSFER_INDIA"
                        ]
                    },
                    "filingYear": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "employee": {
                        "$ref": "#/components/schemas/Employee"
                    },
                    "lca": {
                        "$ref": "#/components/schemas/LCA"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Wage": {
                "type": "object",
                "properties": {
                    "wageId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "wage": {
                        "type": "number",
                        "format": "float"
                    },
                    "wageType": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "wageNotes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Notes"
                        }
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
...
            "PotentialEmployee": {
                "type": "object",
                "properties": {
                    "peId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "emailId": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "dob": {
                        "type": "string"
                    },
                    "company": {
                        "type": "string"
                    },
                    "visaType": {
                        "type": "string"
                    },
                    "referredBy": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "year": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "type": "string"
                    },
                    "primarySkills": {
                        "type": "string"
                    },
                    "secondarySkills": {
                        "type": "string"
                    },
                    "currentLocation": {
                        "type": "string"
                    },
                    "workCountry": {
                        "type": "string"
                    },
                    "onProject": {
                        "type": "string"
                    },
                    "peNotes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Notes"
                        }
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
...
        }
    }
}